Wrap secretstream init function and state struct
Define a foreign struct type :crypto-secretstream-state as a counterpart for
crypto_secretstream_xchacha20poly1305_state and implement wrapper & binding to
the init function that generates the header.

Custom defined foreign types, including structs, can apparently not be passed by
value from c-inline et cetera. Passing by void-pointer would require manual
memory management and additional functions for it. The struct is therefore
passed to the wrapper as a lisp object of type t_foreign (struct ecl_foreign).
The wrapper then passes a pointer to the struct data to libsodium. Memory is
managed by the lisp runtime through ffi:with-foreign-object.

The struct members' sizes are taken from libsodium source. The corresponding
preprocessor constants and C functions can unfortunately not be evaluated to
produce their values at lisp compile-time and need to be hardcoded instead. The
definitions for the struct and the size constants can be found in the following
files:

https://github.com/jedisct1/libsodium/blob/master/src/libsodium/include/sodium/crypto_secretstream_xchacha20poly1305.h

https://github.com/jedisct1/libsodium/blob/master/src/libsodium/include/sodium/crypto_aead_xchacha20poly1305.h