Wrap & bind secretstream push function
Define a foreign type :unsigned-long-long to wrap C's analogous type. The ECL
Manual suggests that :long-long is capable of holding the latter, but a distinct
type denotes the variables' contents more clearly.

Write a C wrapper for crypto_secretstream_xchacha20poly1305_push that translates
from lisp vector types to c pointers, calls secretstream's push function and
returns the length of the resulting ciphertext.

Write a lisp binding that determines lengths if necessary and calls the wrapper.