Define fread-bytes
fread returns the number of complete elements read. If elements are larger than
size 1 and the length of the data read is not a multiple of the element size, it
is not possible to know how much data has been read.

fread-bytes attempts to fill its vector argument with elements of size 1 and
returns the number of bytes read.