Define assert/length macro and corresponding condition
assert-size-not-exceeds-sequence-length expects two expressions: A sequence and
a number. If the number is in the interval between 0 and the length of the given
sequence, the expression it expands to produces the value of the number. If it
is not part of the interval, a size-exceeds-sequence-length-error is signaled
and a restart that produces the sequence length is provided.

assert/length is defined as a shorthand for the unwieldy long name.

The function cap-size-to-sequence-length emits a warning that informs about the
expressions involved in the failed assertion and the length value that will be
substituted, then attempts to invoke the restart.

The main operation execution is wrapped to bind the described restart to any
such errors which might occur.

So far it could be argued that the same effect could have been achieved with a
macro that simply substitutes the length if the assertion fails, without the
need to define a condition, a restart, and a handler. However it can be imagined
that in certain cases, operations will depend on an exact length or part of a
sequence. In these cases it might not be possible to successfully continue
operation by simply substituting the length of the sequence.

The fread and fwrite bindings as well as the bindings for
secretstream-push/-pull are updated to use the new macro to avoid accessing
protected memory and resulting segmentation faults when accessing data via
pointers in C code.

The macro with-gensyms is added to provide a simple tool to avoid
unintentionally evaluating argument expressions multiple times.